inicio

+++ database/migrations/2020_10_09_173929_create_usuario_rol_table.php
@@ -14,9 +14,9 @@ class CreateUsuarioRolTable extends Migration
     public function up()
     {
         Schema::create('usuario_rol', function (Blueprint $table) {
-            $table->unsignedBigInteger('rol_id');
+            $table->unsignedInteger('rol_id');
             $table->foreign('rol_id', 'fk_usuariorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
-            $table->unsignedBigInteger('usuario_id');
+            $table->unsignedInteger('usuario_id');
             $table->foreign('usuario_id', 'fk_usuariorol_usuario')->references('id')->on('usuario')->onDelete('restrict')->onUpdate('restrict');
             $table->timestamps();
         });
